+++ spec/inputs/file_spec.rb
@@ -5,9 +5,15 @@ require "tempfile"
 require "stud/temporary"
 require "logstash/inputs/file"
 
-Thread.abort_on_exception = true
-
 describe "inputs/file" do
+  before(:all) do
+    @abort_on_exception = Thread.abort_on_exception
+    Thread.abort_on_exception = true
+  end
+
+  after(:all) do
+    Thread.abort_on_exception = @abort_on_exception
+  end
 
   delimiter = (LogStash::Environment.windows? ? "\r\n" : "\n")
 
@@ -183,7 +189,7 @@ describe "inputs/file" do
     end
   end
 
-  context "when #run is called multiple times" do
+  context "when #run is called multiple times", :unix => true do
     let(:tmpdir_path)  { Stud::Temporary.directory }
     let(:sincedb_path) { Stud::Temporary.pathname }
     let(:file_path)    { "#{tmpdir_path}/a.log" }
@@ -194,7 +200,7 @@ describe "inputs/file" do
         Thread.new(input, arr) do |i, a|
           sleep 0.5
           a << `lsof -p #{Process.pid} | grep "a.log"`
-          i.stop
+          i.teardown
         end
       end
     end
@@ -221,71 +227,6 @@ describe "inputs/file" do
       stop_proc.call(subject, lsof)
       subject.run(buffer)
       expect(lsof.last).to eq(lsof.first)
-    end
-  end
-
-  context "when wildcard path and a multiline codec is specified" do
-    let(:tmpdir_path)  { Stud::Temporary.directory }
-    let(:sincedb_path) { Stud::Temporary.pathname }
-    let(:conf) do
-      <<-CONFIG
-        input {
-          file {
-            type => "blah"
-            path => "#{tmpdir_path}/*.log"
-            start_position => "beginning"
-            sincedb_path => "#{sincedb_path}"
-            delimiter => "#{FILE_DELIMITER}"
-            codec => multiline { pattern => "^\s" what => previous }
-          }
-        }
-      CONFIG
-    end
-
-    let(:writer_proc) do
-      -> do
-        File.open("#{tmpdir_path}/a.log", "a") do |fd|
-          fd.puts("line1.1-of-a")
-          fd.puts("  line1.2-of-a")
-          fd.puts("  line1.3-of-a")
-          fd.puts("line2.1-of-a")
-        end
-
-        File.open("#{tmpdir_path}/z.log", "a") do |fd|
-          fd.puts("line1.1-of-z")
-          fd.puts("  line1.2-of-z")
-          fd.puts("  line1.3-of-z")
-          fd.puts("line2.1-of-z")
-        end
-      end
-    end
-
-    after do
-      FileUtils.rm_rf(tmpdir_path)
-    end
-
-    let(:event_count) { 2 }
-
-    it "collects separate multiple line events from each file" do
-      writer_proc.call
-
-      events = input(conf) do |pipeline, queue|
-        queue.size.times.collect { queue.pop }
-      end
-
-      expect(events.size).to eq(event_count)
-
-      e1_message = events[0]["message"]
-      e2_message = events[1]["message"]
-
-      # can't assume File A will be read first
-      if e1_message.start_with?('line1.1-of-z')
-        expect(e1_message).to eq("line1.1-of-z#{FILE_DELIMITER}  line1.2-of-z#{FILE_DELIMITER}  line1.3-of-z")
-        expect(e2_message).to eq("line1.1-of-a#{FILE_DELIMITER}  line1.2-of-a#{FILE_DELIMITER}  line1.3-of-a")
-      else
-        expect(e1_message).to eq("line1.1-of-a#{FILE_DELIMITER}  line1.2-of-a#{FILE_DELIMITER}  line1.3-of-a")
-        expect(e2_message).to eq("line1.1-of-z#{FILE_DELIMITER}  line1.2-of-z#{FILE_DELIMITER}  line1.3-of-z")
-      end
     end
   end
 end